requirements.txt
blessings==1.7
certifi==2019.9.11
chardet==3.0.4
Click==7.0
curtsies==0.3.0
Flask==1.0.4
Flask-Cors==3.0.8
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pygments==2.3.1
pyserial==3.4
requests==2.21.0
six==1.13.0
typing==3.7.4.1
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.16.0
